refactor: providers bridge out of iframe
